Feat: static method displayAccount - Excercise 3.15
'displayAccount' is usefull because it is called several times and
it reduces the code inside the main method.
Static method 'displayAccount' displays the name of the account
owner and the balance of the account. Beacuse this method is also
static and placed the same way inside the driver class 'AccountTest' as the
main method. The compiler is able to find method calls to display
from the main method, even if no object is build from the class
'AccountTest'.