Feat: Solving class-average problem using sentinel-controlled iteration.
What: What is different comparing with counter controlled iteration?
The counter variable is set to 0 intead of one. Before the while
loop is reached the program reads the grade or sentinel value. The
sentinel value (signal value, dummy value or flag value) is a atipcal
value compared to the grade values entered (here -1). Inside the
while-loop-body after the running total is added and the counter is
increased, the program aks the user for a new grade. This is done in
the end inside the while loop so the condition statement statement to
control the program's flow of control has a new value to check for.
Either the loop continues or control is handed to the next statement
after the while loop.