Feat: Nested if...else control statement in counter-controlled iteration
The nested if else statement is inside the body of the while loop.
The counters in the if...else selection statement are there to
increase the respective counters. These counters grasp the
passed and failed examination results. The results are printed
after the user input is requested from the user.

Goal: Learn Java.