Feat: Prefix increment and postfix increment operators
When incrementing or decrementing a variable in a statement by itself,
the prefix increment and postfix increment forms have the same effect,
and the prefix decrement and postfix decrement forms have the same
effect. It's only when a variable appears in the context of a larger
expression that preincrementing and postincrementing the variable have
different effects (and similarly for predecrementing and postdecrementing).

Goal: Learn Java