Feat: GUI and Graphics - Event Handling
JavaFx GUI are control objects for interaction. Instance variables
and methods are declared in a class. For an FXML GUI these are
declared in controller class. The controller class controls the
app's acctions in response to the user's interactions with the
GUI.
The JavaFx runtime of a JavaFX app with an FXML GUI creates
the GUI.
The JavaFx runtime of a JavaFX app with an FXML GUI creates
an object fo the controller class that is specified in the
'Scene Builder' (program).
The JavaFx runtinme of a JavaFX app with an FXML GUI initializes
any of the controller object's instance variables that are
specified in the 'Scene Builder', so that they refer to the
corresponding objects.
The JavaFx runtime of a JavaFX app with an FXML GUI configures
any of the controllers object's event handlers that are specified
in the 'Scene Builder', so that they'll be called when the user
interacts with the corresponding controls(s).

In Scene Builder:
- Specify the instance variable with the 'fx:id' property.
(JavaFX initializes the instance variables)
- A method that is called in responce to a user interaction is
called 'event handler'. This particualar method is called by
the JavaFX runtime on an object of the controller class.

The myFXnameController(.java) class is created with 'Scene Builder'.
It has the same name as the myFXname.fxml file name. The
third file has the name 'myFxname.java'.
The main application class is 'myFXname.java' and loads the FXML
file.

Notes:
The '@FXML' annotation in the class file 'myFXnameController.java'
preceding the instance variable indicates that variable name can
be used in the GUI's FXML file beacause it has an 'fx:id'. The
variable is not explicitly initialized through the programmer.
When the app executes the JavaFx runtime intializes the instance
variables with the SceneBuilders object from the myFXname.fxml
file.
The '@FXML' preceding the the method indicates that this method
can be used in the FXML file to specify a control's event handler.
In this particualar case - When the DrawLines app executes, it
configures drawLinesButtonPressed to execute when the user presses
the DrawLines Button. A Button press generates an ActionEvent
object that contains information about the event that occurred.
The corresponding event-handling method must return void and
receive one ActionEvent parameter. (Here parameter not used)

Notes On The Main Application Class:
The App's main-application class that begins the program's
execution, creates the GUI from the FXML file, creates and
configures the controller-class object and displays the GUI
in a window. The main application class has the following
name structure 'myFXname.java', in this case 'DrawLines.java'.
Every JavaFX app present here (in the book) has three code changes.
1) The class name ('myFXname.java'), Line 9. The FXML filename,
Line 14 and text that is displayed in the application's title
bar, Line 17.