Feat: Salary Calculator - nested selection statement in iteration
The salary calculator calculates the gross pay of a worker. The input
is the hourly salary and the amount of hours worked in a week. Since
several gross pays can be calculated a sentinel-controlled iteration
is used. Nested in the iteration statement is the selection statement
which calculates with a different additionl expression (in the else
statement) if the worker works more than 40 hours in a week.

Goal: Learn Java - nested selection statement in sentinel-controlled
	  iteration statement