Feat: Program finds the two largest numbers of ten
Porgram finds the two largest numbers of ten values entered. If the
actual largest is less than the input, the input value will become
the new largest. The old largest becomes the second largest.
If the input is bigger than the actual second largest and if the
input is also at the same time smaller than the largest then the
input will become the second largest value.
The selection statements are nested in an iteration statement.
Here an iteration-controlled loop is used.

Goal: Learn Java - Nested selections statements, sequence statements
      nested in an iteration statement