Feat: Nested selection and sentinel-controlled iteration in iteration
This program analysis the results of exams. The outer structure of
the program is made of an iteration controlled statement. In this
iteration statement there are a few nested statements, an sentinel-
controlled iteration statement to prompt the user for correct input.
The test condition is outsourced in a static function which tests
the input to give a boolean 'true' to keep on prompting the user to
give a correct input or to jump out of the sentinel-controlled
iteration giving a boolean 'false'.
For more information of the program please see in chapter 4 the
program 'Analysis.java'.

Goal: Learn Java - Prompt the user for correct input.