Feat: Excercise 4.30: Square of Asterisk - Nested Statements
Outer iteration is responsible for the rows. The inner iteration is
responsible for the columns. The outer iteration statement is in
charge of printing a new line at the end of each row ant to increase
the row counter 'row'. Since not all rows are the same in each column
some selection statements are required for printing the asteriks.
There are two selection statements which contain a similar body.
These two selection statements contain an equality expression. The
first boolean expression in the first nested selection statement tests
for the first row and the second boolean expression in the second
selection statement tests for the second row. In their body of the
selection statement they contain a while loop (iteration statement).
In the body of the while loop there are again nested two expressions.
First, a asterisk print statement and a column counter increase
expression (column++).
The third selection statement in the outer while loop (iteration
statement) wants to print the 'vertical' border of the square. So for
this reason the first and last row have to be excluded. So printing
the 'vertical border' has to exclude the first and last row through
the selection statements whose test expresion test for two inequality
conditions. First for not being the first row (row != 1), and second
for row not beeing the last row (row != size). Since both cases
have to be excluded the selection statements have to be coupled with
each other through nesting (logical AND). Inside the body of the
'coupled body selection statement' the printing of the columns in the
row have to be differentiated through selection. An iteration loop
is used to go through each column. But for each row the column
printing is selected to be different dependent of the column counter
value. For the first and last column an asterisk is printed. For all
other columns a blank space character is printed.

Goal: Learn Java - The body of the outer iteration loop selects for
the upper and bottom border rows, and hollow border rows. The upper
and bottom border rows print in the column dimension for each column
an asterisk. The hollow border row prints only for the first and last
columns an asteriks, for all the rest column values a blank space
character is printed.

*** Mulitiple nested Iteration and Selection Statements ***