Feat: Excercise 4.31: Test for Palindrome Number
The program tests if a number is a palindrome. The program is divided
in three tasks. The first task is to prompt the user so long to enter
a five digit number until the number is entered. The second task
is to split the number into five digits. The last task is to test
the digits of the number entered if the number is a palindrome.
The methods used here are the iteration and selection statements.
The first task is the most interesting. Here the nested iteration
statement in another iteration statement should take the focus.
This while loop is interesting because it has two variables in the
same nested while loop that change their values through each
iteration. These to variables, the 'denominator' and the 'number'
are initialized differently. The 'denomintor' is initialized with
a '1'. The digits counter 'digits' is initialized with a '0'.
The 'digits' counter is increased with one through each iteration.
The 'denominator' is mulipied by ten through each iteration. The
nested while loop starts by testing its boolean expression if number
entered by the user divided by the denominator is not equal to zero.
Because if the inverted equality expression is true this means the
denominator has not yet reache a bigger the decimal position
compared to the number entered.