Feat: Excercise 4.31: Test for Palindrome Number
The program tests if a number is a palindrome. The program is divided
in three tasks. The first task is to prompt the user so long to enter
a five digit number until the number is entered correctly. The second
task is to split the number into five digits. The last task is to
test the digits of the number entered for being a palindrome.
The methods used here are the iteration and selection statements.
The first task is the most interesting. Here the nested iteration
statement in outer iteration statement should take the focus.
This while loop is interesting because it has two variables in the
same nested while loop that change their values through each
iteration. These to variables, the 'denominator' and the 'number'
are initialized differently. The 'denomintor' is initialized with
a '1'. The digits counter 'digits' is initialized with a '0'.
The 'digits' counter is increased with one through each iteration.
The 'denominator' is mulipied by ten through each iteration. The
nested while loop starts by testing its boolean expression. The
expression tests if number entered by the user divided by the
denominator is not equal to zero yet. Because if the whole inequality
expression is true means that the denominator has not already reached
yet a bigger the decimal position compared to the number entered,
so that the digits counter does need to keep incrementing.
The boolean expression becomes false when division calculation
is equal to zero and the while loop stops executing.
The outer loop keeps prompting the user until a five digit number
is entered. Before the nested iteration the variables 'denominator'
and 'digits' have to be set to their initialization state, other-
wise after entering a number with more than five digits the
program won't stop prompting even if a correct number is enterd.
The second task is solve with sequential statements using the
division and remainder operator. The las task is solved with
nested selection statements (AND).

Goal: Learn Java - Keep prompting until number with correct digit
      size is entered.