Feat: Excercise 4.32: Convert Binary Number into Decimal Value
This program converts the binary number input into it's corresponding
decimal value. The program consits of two blocks and a funciton which
calculates the power of two. The power of two is a shifted calculation,
because the first calcuation 2 to the power of zero is one. From then
on the running total of power of 2 is multiplied with the factor 2.
Because of this shifting the counter of the counter-controlled
iteration statement in the function the 'powerOfTwo' is set to zero.
And in the relational expression the counter is SMALLER and not
SMALLER and EQUAL. In the body of the function the running total
is multiplied with factor '2' through each iteration. '2' to the power of
zero is one. This coencides with the 'powerOfTwo' function, since
the running total is initialized with '1' and the while body is never
entered.
The first block calculates the number of digits need in the second
block. The second block combines the decimal value of the power of
two and multiplies it with '1' if the bit is set. The iteration
statement takes the digits from the right to the left.

Goal: Learn Java - Shifted-Counter-Controlled Iteration Statements