Feat: Program checks sides if they could belong to a triangle
This program prompts the user to input three sides. Then it checks if
they could belong to a triangle. The sum of the smallest and second
largest side should (together) be larger than the largest side to
possible all lengths belong to a triangle.

Goal Learn Java - Solve a Problem