Feat: Phytagors equality expression to test sides of a right triangle       Bug File: Excercise_4_36 as.java 4.37
The former program which tested sides of a possibly triangle contains
a bug. If the biggest (largest) side is given in the end the program
gives a wrong answer. This bug is corrected. The second change is that
this program test the pythagoras equality expression. The variables
were changed from double to integers.

Goal: Learn Java - Solve a Problem.