Feat: Excercise 4.38 b) - Caculates the constant of e
File: Excercise_4_38.java

The factorial function is outsourced in a public static function.
The program prompts the user to define the number of addition terms
for calculating the e constant. The e iteration statement contains
a running total variable. The e constant is of type double , so
the type cast operator is used.

Goal: Learn Java - Iteration Statement with decreasing counter.