Feat: Excercise 4.38 c) Calculates the value of e_x
File: Excercise_4_38.java

Program calculates the value of e_x with the number of addition
terms entered by the user. This program has two static functions
outside the main program. The first static function 'factorial'
calculates the factorial of the integer given. The second static
function 'toThePower' recieves a floating-point number and an
integer that represent the power. The body of the function
contains a selection statement for the special case. And for all
other cases a running total to calculate x to the power of the
counter. The e_x expression statement is an counter-controlled
iteration statement. This iteration statement has the
characteristic of a shifted iteration statement.

Goal: Learn Java - Two static functions outside main and at least
      one 'shifted' counter-controlled iteration statement in the
      main.